+++ tests/test_inference.py
@@ -23,10 +23,13 @@
 """
 Test cases for the python API for tsdate.
 """
+import logging
+
 import msprime
 import numpy as np
 import pytest
 import tsinfer
+import tskit
 import utility_functions
 
 import tsdate
@@ -442,3 +445,92 @@ class TestVariational:
                 method="variational_gamma",
                 max_iterations=-1,
             )
+
+    def test_no_existing_mutation_metadata(self):
+        # Currently only the variational_gamma method embeds mutation metadata
+        ts = tsdate.date(self.ts, mutation_rate=1e-8, method="variational_gamma")
+        for m in ts.mutations():
+            print(m)
+            assert "mn" in m.metadata
+            assert "vr" in m.metadata
+            assert m.metadata["mn"] > 0
+            assert m.metadata["vr"] > 0
+
+    def test_existing_mutation_metadata(self, caplog):
+        tables = self.ts.dump_tables()
+        m = self.ts.mutation(-1)
+        tables.mutations.truncate(self.ts.num_mutations - 1)
+        tables.mutations.metadata_schema = tskit.MetadataSchema.permissive_json()
+        tables.mutations.append(m.replace(metadata={"test": 7}))
+        ts = tables.tree_sequence()
+        with caplog.at_level(logging.WARNING):
+            dts = tsdate.variational_gamma(ts, mutation_rate=1e-8)
+            assert caplog.text == ""
+        assert dts.mutation(-1).metadata["test"] == 7
+        for m in dts.mutations():
+            assert m.metadata["mn"] > 0
+            assert m.metadata["vr"] > 0
+
+    def test_existing_byte_mutation_metadata(self, caplog):
+        tables = self.ts.dump_tables()
+        m = self.ts.mutation(-1)
+        tables.mutations.truncate(self.ts.num_mutations - 1)
+        tables.mutations.append(m.replace(metadata=b"x"))
+        ts = tables.tree_sequence()
+        with caplog.at_level(logging.WARNING):
+            dts = tsdate.variational_gamma(ts, mutation_rate=1e-8)
+            assert "Could not set" in caplog.text
+        assert dts.mutation(-1).metadata == b"x"
+
+    def test_existing_struct_mutation_metadata(self, caplog):
+        tables = self.ts.dump_tables()
+        tables.mutations.metadata_schema = tskit.MetadataSchema(
+            {
+                "codec": "struct",
+                "type": "object",
+                "properties": {
+                    "mn": {
+                        "type": "number",
+                        "binaryFormat": "f",
+                        "description": "Posterior mean mutation time",
+                    },
+                    "vr": {
+                        "type": "number",
+                        "binaryFormat": "f",
+                        "description": "Posterior variance in mutation time",
+                    },
+                },
+                "additionalProperties": False,
+            }
+        )
+        tables.mutations.packset_metadata(
+            [
+                tables.mutations.metadata_schema.validate_and_encode_row(
+                    {"mn": 0, "vr": 0}
+                )
+                for _ in range(tables.mutations.num_rows)
+            ]
+        )
+        ts = tables.tree_sequence()
+        with caplog.at_level(logging.WARNING):
+            dts = tsdate.variational_gamma(ts, mutation_rate=1e-8)
+            assert caplog.text == ""
+        for m in dts.mutations():
+            assert m.metadata["mn"] > 0
+            assert m.metadata["vr"] > 0
+
+    def test_incompatible_schema_mutation_metadata(self, caplog):
+        tables = self.ts.dump_tables()
+        tables.mutations.metadata_schema = tskit.MetadataSchema(
+            {
+                "codec": "struct",
+                "type": "object",
+                "properties": {},
+                "additionalProperties": False,
+            }
+        )
+        ts = tables.tree_sequence()
+        with caplog.at_level(logging.WARNING):
+            dts = tsdate.variational_gamma(ts, mutation_rate=1e-8)
+            assert "Could not set" in caplog.text
+        assert len(dts.tables.mutations.metadata) == 0